+++ osidb/migrations/0223_fix_update_stream.py
@@ -8,11 +8,23 @@ corresponding module (based on product definitions)
 """
 
 
+import re
+
 from django.conf import settings
 from django.db import migrations
 
 from osidb.core import set_user_acls
-from osidb.helpers import ps_update_stream_natural_keys
+
+
+def ps_update_stream_natural_keys(values):
+    """Sort alphanumeric strings (natural/human sorting)"""
+    if not values:
+        return []
+
+    def atoi(text):
+        return int(text) if text.isdigit() else text
+
+    return [atoi(c) for c in re.split(r"(\d+)", values.name)]
 
 
 PS_UPDATE_STREAM_LIST = [